+++ Releases/Utility Mod/Changelog.md
@@ -1,5 +1,6 @@
 ## v2.4.2 (unreleased)
 - Allow TankBro's tank to spawn when enemy spawns are disabled.
+- Added fix for reversed and slow scroll wheel in mod manager UI on Linux.
 
 ## v2.4.1
 - Added option to go to custom campaigns from right click menu.